+++ usb-hid-s3/test/test_device_identity/test_device_identity.cpp
@@ -13,10 +13,10 @@ void test_device_id_from_mac() {
   TEST_ASSERT_EQUAL_STRING("eeff", suffix);
 }
 
-void test_soft_ap_ssid_upper_suffix() {
+void test_soft_ap_ssid_uses_friendly_mac_name() {
   char ssid[24];
-  formatSoftApSsid("eeff", ssid);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-EEFF", ssid);
+  formatSoftApSsid("de94", ssid);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Dupe9", ssid);
 }
 
 void test_mdns_hostname_lower_suffix() {
@@ -25,10 +25,19 @@ void test_mdns_hostname_lower_suffix() {
   TEST_ASSERT_EQUAL_STRING("inputpilot-eeff", host);
 }
 
-void test_device_name_uses_upper_suffix() {
+void test_device_name_matches_soft_ap_name() {
   char name[24];
-  formatDeviceName("eeff", name);
-  TEST_ASSERT_EQUAL_STRING("InputPilot-EEFF", name);
+  formatDeviceName("DE94", name);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Dupe9", name);
+}
+
+void test_friendly_name_has_deterministic_fallbacks() {
+  char digitsOnly[24];
+  char lettersOnly[24];
+  formatDeviceName("1234", digitsOnly);
+  formatDeviceName("beef", lettersOnly);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Bolt1", digitsOnly);
+  TEST_ASSERT_EQUAL_STRING("InputPilot-Bolt1", lettersOnly);
 }
 
 void test_null_mac_rejected() {
@@ -43,9 +52,10 @@ void tearDown() {}
 int main() {
   UNITY_BEGIN();
   RUN_TEST(test_device_id_from_mac);
-  RUN_TEST(test_soft_ap_ssid_upper_suffix);
+  RUN_TEST(test_soft_ap_ssid_uses_friendly_mac_name);
   RUN_TEST(test_mdns_hostname_lower_suffix);
-  RUN_TEST(test_device_name_uses_upper_suffix);
+  RUN_TEST(test_device_name_matches_soft_ap_name);
+  RUN_TEST(test_friendly_name_has_deterministic_fallbacks);
   RUN_TEST(test_null_mac_rejected);
   return UNITY_END();
 }